requirements.txt
pygame
pynput
pystray
pillow